Result of multiple linear regression model
Note that: I used all the key variables identified by James and Rui to fit the model.
However, I didn't fit the interaction term, because from the residual plot, it does not show much of an improvement of model fitting. And also, it adds great complexity to model interpretation. 
I also fitted the model using a sqrt transformation of the price (it also aligns with our SLR), after fitting this, the model fitting has been greatly improved and we maintain a relatively clean model that's easier to understand.